Feat: Set up delete coworking space.

+++ src/components/DeleteSpaceCwForm.js
@@ -1,15 +1,15 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
+import { useDispatch } from 'react-redux';
 import styles from '../assets/stylesheets/DeleteSpaceCwForm.module.css';
+import { CLEAR_PERSISTED_STATE } from '../actions/clearState';
 
 const DeleteSpaceCwForm = () => {
+  const dispatch = useDispatch();
   const userId = JSON.parse(localStorage.getItem('user'))?.user.id;
 
-  const [formData, setFormData] = useState({
-    space_cw_id: '',
-    user_id: userId,
-  });
-
+  // Initialized with null to ensure no space is unintentionally selected
+  const [selectedSpaceCwId, setSelectedSpaceCwId] = useState(null);
   const [spaceCws, setSpaceCws] = useState([]);
   const [successMessage, setSuccessMessage] = useState('');
   const [errorMessage, setErrorMessage] = useState('');
@@ -18,37 +18,45 @@ const DeleteSpaceCwForm = () => {
   useEffect(() => {
     const fetchSpaceCws = async () => {
       try {
-        const response = await axios.get(`http://localhost:3001/api/v1/users/${formData.user_id}/space_cws`);
+        const response = await axios.get(`http://localhost:3001/api/v1/users/${userId}/space_cws`);
         setSpaceCws(response.data);
       } catch (error) {
         setFetchError('Error fetching coworking spaces');
       }
     };
 
-    if (formData.user_id) {
+    if (userId) {
       fetchSpaceCws();
     }
-  }, [formData.user_id]);
+  }, [userId]);
 
   const handleChange = (e) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value });
-    console.log('Form Data before submission:', formData);
+    setSelectedSpaceCwId(e.target.value);
   };
 
   const handleSubmit = async (e) => {
     e.preventDefault();
+    if (!selectedSpaceCwId) {
+      setErrorMessage('Please select a coworking space.');
+      setTimeout(() => {
+        setErrorMessage('');
+      }, 5000); // Display the error for 5 seconds
+      return;
+    }
+
     try {
-      const delUrl = `http://localhost:3001/api/v1/users/${formData.user_id}/space_cws/${formData.space_cw_id}`;
+      const delUrl = `http://localhost:3001/api/v1/users/${userId}/space_cws/${selectedSpaceCwId}`;
       await axios.delete(delUrl);
       setSuccessMessage('Coworking space was deleted successfully!');
+      dispatch({ type: CLEAR_PERSISTED_STATE });
       setTimeout(() => {
         setSuccessMessage('');
-      }, 5000); // 5 seconds
+      }, 5000); // Hide the message after 5 seconds
     } catch (error) {
       setErrorMessage('Error deleting the coworking space. Please try again later.');
       setTimeout(() => {
         setErrorMessage('');
-      }, 8000); // 8 seconds
+      }, 8000); // Hide the message after 8 seconds
     }
   };
 
@@ -63,11 +71,11 @@ const DeleteSpaceCwForm = () => {
               id="space_cw_id"
               name="space_cw_id"
               className={styles.formField}
-              value={formData.space_cw_id}
+              value={selectedSpaceCwId || ''}
               onChange={handleChange}
               required
             >
-              <option value="">Select a Workspace</option>
+              <option value="" disabled>Select a Workspace</option>
               {spaceCws.map((spaceCw) => (
                 <option key={spaceCw.id} value={spaceCw.id}>
                   {spaceCw.name}
@@ -76,19 +84,16 @@ const DeleteSpaceCwForm = () => {
             </select>
           </label>
 
-          {/* type="hidden"; value={loggedInUserId} */}
-          {/* <label htmlFor="user_id">User Id:</label> */}
           <input
             type="hidden"
             name="user_id"
-            value={formData.user_id}
+            value={userId}
             required
           />
 
-          {/* Success and error messages */}
-          {successMessage && <p className="success-message">{successMessage}</p>}
-          {errorMessage && <p className="error-message">{errorMessage}</p>}
-          {fetchError && <p className="error-message">{fetchError}</p>}
+          {successMessage && <p className={styles.successMessage}>{successMessage}</p>}
+          {errorMessage && <p className={styles.errorMessage}>{errorMessage}</p>}
+          {fetchError && <p className={styles.errorMessage}>{fetchError}</p>}
 
           <button type="submit" className={styles.deleteSpaceBtn}>
             Delete Coworking Space
